MS1/requirements.txt
psycopg2-binary
argon2-cffi
flask
pika

